refactor(map): déplace HexGrid + ariaHexWorld → features/map/
Distinction world/ vs map/ :
- world/ = logique métier (moteur, crises, données pays)
- map/   = rendu visuel (hexagones SVG, géométrie)

HexGrid.jsx : import WorldEngine corrigé → '../world/WorldEngine'
Dashboard_p2.jsx : import HexGrid corrigé → './features/map/HexGrid'
ARBORESCENCE.md : structure mise à jour + règle world/map + dettes techniques
